requirements.txt
ImageHash==4.2.1
numpy==1.21.6
Pillow==9.2.0
PyWavelets==1.3.0
scipy==1.7.3
six==1.16.0
